add a "dinosaur" tag to the oldest entries that exist
The COT JSON objects have a property "dinosaur" if it is the oldest one in its series. For example, if the WTI Phyiscal Crude Oil contract goes back to 2006-01-01, the JSON object corresponding to that date's entry has a "dinosaur" property. By checking that, callers know in advance whether or not to bother making request to the third-party Socrata API. This improves on the whole idea of caching and not making unnecessary requests.